requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
telegramify-markdown==0.5.4
python-dotenv==1.2.1
pydantic==2.12.5
mistletoe==1.4.0
